pattern cleanup: moved mm pattern code to pattern/ dir
