+++ pachi.c
@@ -14,9 +14,9 @@
 #include "engines/replay.h"
 #include "engines/montecarlo.h"
 #include "engines/random.h"
-#include "engines/patternscan.h"
-#include "engines/pattern.h"
 #include "engines/dcnn.h"
+#include "pattern/patternscan_engine.h"
+#include "pattern/pattern_engine.h"
 #include "joseki/joseki_engine.h"
 #include "joseki/josekiscan_engine.h"
 #include "t-unit/test.h"
@@ -32,9 +32,9 @@
 #include "fifo.h"
 #include "dcnn.h"
 #include "caffe.h"
-#include "pattern.h"
-#include "patternsp.h"
-#include "patternprob.h"
+#include "pattern/pattern.h"
+#include "pattern/spatial.h"
+#include "pattern/prob.h"
 #include "joseki/joseki.h"
 
 /* Main options */
