+++ pattern/Makefile
@@ -1,0 +1,8 @@
+INCLUDES=-I..
+
+OBJS := pattern.o  pattern_engine.o  patternscan_engine.o  prob.o  spatial.o
+
+all: lib.a
+lib.a: $(OBJS)
+
+-include ../Makefile.lib
