+++ pattern/pattern.h
