+++ pattern/pattern_engine.h
@@ -2,7 +2,7 @@
 #define PACHI_PATTERN_ENGINE_H
 
 #include "engine.h"
-#include "../pattern.h"
+#include "pattern/pattern.h"
 
 void pattern_engine_init(engine_t *e, board_t *b);
 pattern_config_t *pattern_engine_get_pc(engine_t *e);
