+++ pattern/patternscan_engine.h
