+++ pattern/training/README
