+++ pattern/training/byplayer
