+++ pattern/training/mm/Makefile
