+++ pattern/training/mm/example.dat
