+++ pattern/training/mm/mm.cpp
