+++ pattern/training/mm_gammas
