+++ pattern/training/spatial_gen
