+++ pattern/training/spatial_show
