+++ uct/internal.h
@@ -6,9 +6,9 @@
 #include "debug.h"
 #include "move.h"
 #include "ownermap.h"
-#include "pattern.h"
-#include "patternsp.h"
-#include "patternprob.h"
+#include "pattern/pattern.h"
+#include "pattern/spatial.h"
+#include "pattern/prob.h"
 #include "playout.h"
 #include "stats.h"
 #include "mq.h"
